Added the `version` method to the contracts
Improved the default build target to wasm.
In this way, running `cargo build` will default to the wasm target.

* Renamed and update `rust-toolchain` into `rust-toolchain.toml`
* Updated `.cargo/` information
* `build.sh` now also generates the contract's code documentation. In the future, t his documentation should be hosted.
* Added and updated the `build.sh` and `test.sh` of each contract; also added a `test.sh` for the whole repo.
* Overall documentation and comments improvement.

All contracts have the view method `version`, which require no args, and a module
that facilitates this has been created.

* Added a `version.rs` module on `lib-near-ext/`, so contract projects can use it.
* Updated `build.sh` to loop over the files, and also to always trigger the `build.rs` step for each contract project

Example of a response:

```json
{
  name: 'nearapps-counter',
  semver: '0.1.0',
  git_sha: '2beb5ec70ee2c0490cd0f4964544c998e6badbcc',
  git_dirty: true,
  cargo_features: 'default',
  cargo_profile: 'release',
  rustc_semver: '1.56.1',
  rustc_llvm: '13.0',
  rustc_sha: '59eed8a2aac0230a8b53e89d4e99d55912ba6b35'
}
```

The `name` and `semver` is the project name and version as defined on the `Cargo.toml` file.
The `git_sha` is captured by running `git describe --abbrev=40 --always`, and `git_dirty` is whether the `git status --porcelain -u --no-column` doesn't return an empty output - if it's `false`, then no new files nor modified files are present during the contracts compilation.
`cargo_features` and `cargo_profile` are compilation features and profile as set by the `cargo build` command, in which the `cargo_features` fallbacks to `default` when none are set.
`rustc_semver`, `rustc_llvm` and `rustc_sha` are information of the compiler itself, such as it's version and so on.