Added stateful token for "sending" operations
Tokens that "started to get sent" must change the state to prevent re-entrancy attacks; but the tokens also must not be entirely "removed" from the user because, in case that was the last token for a registered nft contract on the send-nft contract, that nft contract could be de-registered from it - that is, in case this operation happened between that last token sending operation and it's callback execution. In this case, the nft contract could fail to send the token and the send-nft contract would try to re-register the token back to the user in a situation where the nft contract would no longer exist, causing the token to be lost.
Therefore to avoid this and still be safe from re-entrant attacks, the "token sending operation" must set tokens into a different state (but still registered on the user), and then only completely de-register from the user when the sending operation was indicated as successful from the nft contract.

* Added stateful tokens for sending operations.
* Removed the goal of not-panicking on the callback methods, simplifying the code. In those cases, the token will be lost as it will be stuck as "OnSending".

Extra updates:

* Removed default target since it would require manual configuration of rust-analyzer, otherwise it wouldn't work.
* Removed executor dependency for the counter, otherwise they would have a cyclic dependency.
  * The counter then doesn't check for logging, but other apps do.
* Added new functions for the `fun!()` macro on `lib-near-ext`.